+++ src/Trash/Sonarr/QualityDefinition/SonarrQualityDefinitionGuideParser.cs
@@ -17,7 +17,7 @@ namespace Trash.Sonarr.QualityDefinition
         public async Task<string> GetMarkdownData()
         {
             return await
-                "https://raw.githubusercontent.com/TRaSH-/Guides/master/docs/Sonarr/V3/Sonarr-Quality-Settings-File-Size.md"
+                "https://raw.githubusercontent.com/TRaSH-/Guides/master/docs/Sonarr/Sonarr-Quality-Settings-File-Size.md"
                     .GetStringAsync();
         }