+++ CHANGELOG.md
@@ -8,6 +8,12 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+
+- Sonarr: Use new URL for release profile guide.
+- Sonarr: Use new URL for quality definition guide.
+- Radarr: Use new URL for quality definition guide.
+
 ## [1.6.0] - 2021-05-31
 
 ### Added